+++ feature/google/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     implementation project(":common")
     implementation project(":public")
     implementation project(":strings")
+    implementation project(":utils")
 
     implementation "androidx.core:core-ktx:1.3.1"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
